requirements.txt
# core (Python 3.12 compatible)
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4
pillow==10.3.0
scikit-learn==1.4.2
matplotlib==3.8.4
opencv-python==4.9.0.80

# server & monitoring
fastapi==0.110.0
uvicorn[standard]==0.29.0
prometheus-client==0.20.0

# tooling
mlflow==2.12.1
dvc==3.51.1
pytest==8.2.0

# utils
python-dotenv==1.0.1
gunicorn==21.2.0

python-multipart==0.0.9